refactor the query function
