--- components/components-jdbc/components-jdbc-definition/src/main/java/org/talend/components/jdbc/query/generator/PostgreQueryGenerator.java
@@ -1,23 +1,0 @@
-// ============================================================================
-//
-// Copyright (C) 2006-2017 Talend Inc. - www.talend.com
-//
-// This source code is available under agreement available at
-// %InstallDIR%\features\org.talend.rcp.branding.%PRODUCTNAME%\%PRODUCTNAME%license.txt
-//
-// You should have received a copy of the agreement
-// along with this program; if not, write to Talend SA
-// 9 rue Pages 92150 Suresnes, France
-//
-// ============================================================================
-package org.talend.components.jdbc.query.generator;
-
-import org.talend.components.jdbc.query.EDatabaseTypeName;
-
-public class PostgreQueryGenerator extends DefaultQueryGenerator {
-
-    public PostgreQueryGenerator(EDatabaseTypeName dbType) {
-        super(dbType);
-    }
-
-}
